Handle the Rigth Alt key for Bucket 3.
For injection it has to be released as an extended key else it creates
all kinds of issues, as the Right Alt is pressed and we are sending a
key release for the Left Alt which is already released.
This caused Right Alt to not work for Bucket 3 and also caused other
key issues.

Signed-off-by: Suriyan Ramasami <suriyan.r@gmail.com>